+++ tags/reserved_tags.go
@@ -3,11 +3,8 @@ package tags
 import (
 	"strings"
 
-	"github.com/google/martian/v3/tags"
 	"github.com/pkg/errors"
 )
-
-type Key = tags.Key
 
 const (
 	// Identifies the source of a trace or spec. See `Source` for values.
@@ -100,27 +97,4 @@ func CheckReservedKey(k Key) error {
 	}
 
 	return errors.New(`Tags starting with "x-akita-" are reserved for Akita internal use.`)
-}
-
-// Returns a map from parsing a list of "key=value" pairs.
-func FromPairs(pairs []string) (map[Key]string, error) {
-	results := make(map[Key]string, len(pairs))
-	for _, p := range pairs {
-		parts := strings.Split(p, "=")
-		if len(parts) != 2 {
-			return nil, errors.Errorf("%s is not a valid key=value format", p)
-		}
-
-		k, v := Key(parts[0]), parts[1]
-		if _, ok := results[k]; ok {
-			return nil, errors.Errorf("tag with key %s specified more than once", k)
-		}
-
-		if err := CheckReservedKey(k); err != nil {
-			return nil, err
-		}
-
-		results[k] = v
-	}
-	return results, nil
 }
